Now we can upload our file images !
-Add the option to host an image into a cloud and take the image's link and add it in the data base.
-Modify the create, edit and delete routes:
Create: Host the input image in the cloud and give it an imageId and imageURL.
Edit: Change the image of the campground in the data base and erase the last one.
Delete: Remove the comments of the campground, the image from the cloud and the entire campground from the data base.